MS-1 some gurus are premium

+++ uni-scrape/src/main/java/com/tvm/Guru.java
@@ -1,12 +1,15 @@
 package com.tvm;
 
 import com.gargoylesoftware.htmlunit.BrowserVersion;
+import com.gargoylesoftware.htmlunit.ElementNotFoundException;
 import com.gargoylesoftware.htmlunit.WebClient;
 import com.gargoylesoftware.htmlunit.html.*;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
@@ -42,14 +45,25 @@ public class Guru {
         List<?> list = page.getByXPath("//a[@class='gurunames']");
         for(Object obj : list) {
             HtmlAnchor a = (HtmlAnchor)obj;
-            System.out.println(a);
-            page = a.click();
-            a = page.getAnchorByText("Download");
-            System.out.println(a);
-            String fileName = a.getHrefAttribute().replace("/download_StockBuy.php?GuruName=", "");
+
+            String fileName = a.getHrefAttribute().replace("/StockBuy.php?GuruName=", "");
             fileName = fileName + ".csv";
 
-            IOUtils.copy(a.click().getWebResponse().getContentAsStream(), new FileOutputStream(fileName));
+            if(new File(fileName).exists()) {
+                System.out.println(fileName+" exists, continuing");
+                continue;
+            }
+
+            System.out.println(a);
+            page = a.click();
+
+            try {
+                a = page.getAnchorByText("Download");
+                System.out.println(a);
+                IOUtils.copy(a.click().getWebResponse().getContentAsStream(), new FileOutputStream(fileName));
+            } catch(ElementNotFoundException e) {
+                System.out.println("No download link found");
+            }
 
 
         }
